Add warning about missing autotiler file to documentation
Added a warning to documentation regarding the unavailability of the required autotiler file for neural network deployment with gap_sdk due to the GreenWaves Technologies website being down. The warning informs users that examples relying on the autotiler will not work unless they already have the file and provides a link to the announcement with more details and workarounds.